+++ CHANGELOG.md
@@ -5,7 +5,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 **Changes to JS assets (including the front-end and JS libraries) are not shown here**, but in [`/browser/CHANGELOG`](/browser/CHANGELOG.md).
 See [STATUS.md](server/STATUS.md) to learn more about which features will remain stable.
 
-## [v0.38.0] - UNRELEASED
+## [v0.38.0] - 2024-06-08
 
 - Remove `process-management` feature #324 #334
 - Add `atomic_lib::client::search` for building queries #778